몬스터 공격

+++ Client/CScene.h
@@ -83,7 +83,7 @@ public:
 
 	int GetMapRowSize() { return m_WallMap.size(); }
 	int GetMapColSize() { return m_WallMap[0].size(); }
-	int GetWallmapNum(float x, float y); // 해당 좌표 셀의 숫자 반환
+	int GetWallmapNum(int x, int y); // 해당 좌표 셀의 숫자 반환
 
 public:
